Nova camada de busca proximal antes do prompt para o Weaviate
